requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-crontab==0.7.1
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
sqlparse==0.4.4
pytz==2022.1
six==1.16.0
Werkzeug==2.3.6
numpy==1.25.1